# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ANAM-BF - API de la plateforme agrométéo",
    "description": "\nAPI de la plateforme agrométéorologique de l'ANAM-BF.\n\n**Conventions**\n- Format JSON, UTF-8. Dates en UTC au format ISO 8601.\n- Authentification : en-tête `Authorization: Bearer <access_token>` (voir la section *Authentification*).\n  Les contenus publiés (bulletins, alertes, avis, carte, zones) sont **publics** : consultables sans compte.\n- Erreurs : `{\"detail\": \"message lisible\"}` avec le code HTTP adapté (401 non connecté, 403 droits insuffisants,\n  404 introuvable, 409 conflit d'état, 422 données invalides, 429 trop de tentatives).\n- Listes paginées : paramètres `limit` et `offset`, réponse `{\"total\": N, \"items\": [...]}`.\n- Langues : `fr` (français), `en` (anglais), `mos` (mooré).\n- Génération audio/vidéo : asynchrone. Lancer, puis relire le contenu jusqu'à `media.status` = `ready`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/me": {
      "get": {
        "tags": [
          "Profil"
        ],
        "summary": "Get Me",
        "description": "**Accès :** Connexion requise (tout utilisateur connecté).",
        "operationId": "get_me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Profil"
        ],
        "summary": "Update Me",
        "description": "Met à jour la commune, la langue préférée et le mode de réception des alertes.\n\n**Accès :** Connexion requise (tout utilisateur connecté).",
        "operationId": "update_me_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me/password": {
      "post": {
        "tags": [
          "Profil"
        ],
        "summary": "Change Password",
        "description": "Change le mot de passe et ferme toutes les sessions (reconnexion nécessaire).\n\n**Accès :** Connexion requise (tout utilisateur connecté).",
        "operationId": "change_password_me_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/media/{path}": {
      "get": {
        "tags": [
          "Plateforme",
          "Médias"
        ],
        "summary": "Get Media",
        "description": "Fichier généré (audio, vidéo, image, PDF). Les médias des contenus **publiés** sont publics ; ceux\ndes brouillons ne sont lisibles que par le personnel (en-tête `Authorization`, ou `?token=`).\nLes URLs à utiliser sont fournies dans les champs `media` / `image_url` / `pdf_url` des contenus.\n\n**Accès :** Public. Une connexion avec droits de gestion donne accès aux contenus non publiés.",
        "operationId": "get_media_media__path__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Token d'accès, pour lire un média non publié dans une balise <video>/<audio>",
              "title": "Token"
            },
            "description": "Token d'accès, pour lire un média non publié dans une balise <video>/<audio>"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "post": {
        "tags": [
          "Plateforme",
          "Statistiques d'usage"
        ],
        "summary": "Track Event",
        "description": "À appeler par l'application quand un contenu est consulté (`view`), écouté (`play_audio`), regardé\n(`play_video`) ou partagé (`share`) : alimente les statistiques d'usage du back-office. Connexion facultative.\n\n**Accès :** Public. Une connexion avec droits de gestion donne accès aux contenus non publiés.",
        "operationId": "track_event_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me/devices": {
      "post": {
        "tags": [
          "Plateforme",
          "Profil"
        ],
        "summary": "Register Device",
        "description": "Enregistre le jeton Firebase (FCM) de l'appareil pour recevoir les notifications push.\nÀ appeler après chaque connexion et quand Firebase renouvelle le jeton.\n\n**Accès :** Connexion requise (tout utilisateur connecté).",
        "operationId": "register_device_me_devices_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegister"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Plateforme",
          "Profil"
        ],
        "summary": "Unregister Device",
        "description": "Supprime le jeton de l'appareil (à appeler à la déconnexion).\n\n**Accès :** Connexion requise (tout utilisateur connecté).",
        "operationId": "unregister_device_me_devices_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 10,
              "title": "Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Channel": {
        "type": "string",
        "enum": [
          "sms",
          "push",
          "email"
        ],
        "title": "Channel"
      },
...
      "DeviceRegister": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 10,
            "title": "Token",
            "description": "Jeton Firebase (FCM) de l'appareil"
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios",
              "web"
            ],
            "title": "Platform",
            "default": "android"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "DeviceRegister"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Language": {
        "type": "string",
        "enum": [
          "fr",
          "mos",
          "en"
        ],
        "title": "Language"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
      "PasswordChange": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChange"
      },
...
      "ProfileUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "commune": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune"
          },
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Language"
              },
              {
                "type": "null"
              }
            ]
          },
          "notification_channels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Channel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Channels"
          }
        },
        "type": "object",
        "title": "ProfileUpdate"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "grand_public",
          "responsable_communal",
          "agent_anam",
          "observateur",
          "administrateur"
        ],
        "title": "Role"
      },
...
      "UsageEventIn": {
        "properties": {
          "content_type": {
            "type": "string",
            "enum": [
              "alert",
              "bulletin",
              "advisory"
            ],
            "title": "Content Type"
          },
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "view",
              "play_audio",
              "play_video",
              "share"
            ],
            "title": "Kind",
            "default": "view"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fr",
                  "en",
                  "mos"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "content_id"
        ],
        "title": "UsageEventIn"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "commune": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "notification_channels": {
            "items": {
              "$ref": "#/components/schemas/Channel"
            },
            "type": "array",
            "title": "Notification Channels"
          },
          "totp_enabled": {
            "type": "boolean",
            "title": "Totp Enabled"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "phone",
          "email",
          "username",
          "full_name",
          "role",
          "status",
          "commune",
          "language",
          "notification_channels",
          "totp_enabled",
          "must_change_password",
          "created_at",
          "last_login_at"
        ],
        "title": "UserOut"
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "disabled"
        ],
        "title": "UserStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}